Enables use of `_functionBuilder`
Function Builders are still pretty beta (note the underscore), and they don't play nice with closures with only a single entry (which triggers implicit return behavoir, instead). So we had to refactor our builders to expect an abstract `FieldSerializable` so that the implict return case would still be valid.

Ultimately this feels like a more consistent design (and our builder is just passing along its arguments, so short-circuiting around it isn't a bug). So I'm comfortable moving forward with it even in the face of changing behaviors in future Swifts.